Disable check for the heading of the plane
The heading from FlightGear appears to be incorrect. I spent a day
debugging this with the assumption that my calculations were wrong. But
I have verified them from several perspectives. It's not impossible that
my calculations are wrong, but it's unlikely. I'm therefore trying not
to rely on heading information until I have a method of cleaning the
heading data.